kloud/kloud: refactor main package
