+++ go/src/koding/kites/kloud/kloud/kloud_test.go
@@ -1,0 +1,6 @@
+package kloud_test
+
+import "testing"
+
+func TestKloud(t *testing.T) {
+}
